+++ plugins/CoreHome/vue/src/SiteSelector/Site.ts
@@ -6,10 +6,18 @@
  */
 
 interface Site {
-  idsite: string;
+  idsite: string|number;
   name: string;
+  type: string;
   group?: string;
+  timezone: string;
   currency?: string;
+  timezone_name: string;
+  currency_name?: string;
+  alias_urls: string[];
+  excluded_ips: string;
+  excluded_parameters: string;
+  excluded_user_agents: string;
 }
 
 export default Site;