+++ plugins/CorePluginsAdmin/vue/src/index.ts
@@ -15,6 +15,8 @@ import './Form/Form.adapter';
 
 export { default as FormField } from './FormField/FormField.vue';
 export { default as Field } from './Field/Field.vue';
+export { default as Setting } from './PluginSettings/Setting';
+export { default as SettingsForSinglePlugin } from './PluginSettings/SettingsForSinglePlugin';
 export { default as PluginSettings } from './PluginSettings/PluginSettings.vue';
 export { default as PluginFilter } from './Plugins/PluginFilter';
 export { default as PluginManagement } from './Plugins/PluginManagement';